refactor more validation
