+++ convex/_generated/api.d.ts
@@ -183,12 +183,13 @@ import type * as sidebarItems_triggerTypes from "../sidebarItems/triggerTypes.js
 import type * as sidebarItems_triggers from "../sidebarItems/triggers.js";
 import type * as sidebarItems_types_baseTypes from "../sidebarItems/types/baseTypes.js";
 import type * as sidebarItems_types_types from "../sidebarItems/types/types.js";
+import type * as sidebarItems_validation_access from "../sidebarItems/validation/access.js";
 import type * as sidebarItems_validation_color from "../sidebarItems/validation/color.js";
 import type * as sidebarItems_validation_icon from "../sidebarItems/validation/icon.js";
 import type * as sidebarItems_validation_name from "../sidebarItems/validation/name.js";
+import type * as sidebarItems_validation_orchestration from "../sidebarItems/validation/orchestration.js";
 import type * as sidebarItems_validation_parent from "../sidebarItems/validation/parent.js";
 import type * as sidebarItems_validation_slug from "../sidebarItems/validation/slug.js";
-import type * as sidebarItems_validation_validation from "../sidebarItems/validation/validation.js";
 import type * as sidebarShares_functions_getSidebarItemShares from "../sidebarShares/functions/getSidebarItemShares.js";
 import type * as sidebarShares_functions_getSidebarItemSharesForItem from "../sidebarShares/functions/getSidebarItemSharesForItem.js";
 import type * as sidebarShares_functions_getSidebarItemWithShares from "../sidebarShares/functions/getSidebarItemWithShares.js";
@@ -416,12 +417,13 @@ declare const fullApi: ApiFromModules<{
   "sidebarItems/triggers": typeof sidebarItems_triggers;
   "sidebarItems/types/baseTypes": typeof sidebarItems_types_baseTypes;
   "sidebarItems/types/types": typeof sidebarItems_types_types;
+  "sidebarItems/validation/access": typeof sidebarItems_validation_access;
   "sidebarItems/validation/color": typeof sidebarItems_validation_color;
   "sidebarItems/validation/icon": typeof sidebarItems_validation_icon;
   "sidebarItems/validation/name": typeof sidebarItems_validation_name;
+  "sidebarItems/validation/orchestration": typeof sidebarItems_validation_orchestration;
   "sidebarItems/validation/parent": typeof sidebarItems_validation_parent;
   "sidebarItems/validation/slug": typeof sidebarItems_validation_slug;
-  "sidebarItems/validation/validation": typeof sidebarItems_validation_validation;
   "sidebarShares/functions/getSidebarItemShares": typeof sidebarShares_functions_getSidebarItemShares;
   "sidebarShares/functions/getSidebarItemSharesForItem": typeof sidebarShares_functions_getSidebarItemSharesForItem;
   "sidebarShares/functions/getSidebarItemWithShares": typeof sidebarShares_functions_getSidebarItemWithShares;
