+++ convex/blocks/functions/getBlocksWithShares.ts
@@ -4,7 +4,7 @@ import { CAMPAIGN_MEMBER_ROLE } from '../../campaigns/types'
 import { getCampaignMembers } from '../../campaigns/functions/getCampaignMembers'
 import { PERMISSION_LEVEL } from '../../permissions/types'
 import { SHARE_STATUS } from '../../blockShares/types'
-import { checkItemAccess } from '../../sidebarItems/validation/validation'
+import { checkItemAccess } from '../../sidebarItems/validation/access'
 import { findBlockByBlockNoteId } from './findBlockByBlockNoteId'
 import { getSidebarItem } from '../../sidebarItems/functions/getSidebarItem'
 import type { ShareStatus } from '../../blockShares/types'
