+++ convex/bookmarks/functions/toggleItemBookmark.ts
@@ -1,5 +1,5 @@
 import { ERROR_CODE, throwClientError } from '../../errors'
-import { checkItemAccess } from '../../sidebarItems/validation/validation'
+import { checkItemAccess } from '../../sidebarItems/validation/access'
 import { PERMISSION_LEVEL } from '../../permissions/types'
 import { getSidebarItem } from '../../sidebarItems/functions/getSidebarItem'
 import type { CampaignMutationCtx } from '../../functions'
