+++ convex/documentSnapshots/functions/getSnapshot.ts
@@ -1,6 +1,6 @@
 import { ERROR_CODE, throwClientError } from '../../errors'
 import { PERMISSION_LEVEL } from '../../permissions/types'
-import { requireItemAccess } from '../../sidebarItems/validation/validation'
+import { requireItemAccess } from '../../sidebarItems/validation/access'
 import { getSidebarItem } from '../../sidebarItems/functions/getSidebarItem'
 import type { CampaignQueryCtx } from '../../functions'
 import type { Id } from '../../_generated/dataModel'
