+++ convex/gameMaps/functions/createItemPin.ts
@@ -1,4 +1,4 @@
-import { requireItemAccess } from '../../sidebarItems/validation/validation'
+import { requireItemAccess } from '../../sidebarItems/validation/access'
 import { getSidebarItem } from '../../sidebarItems/functions/getSidebarItem'
 import { PERMISSION_LEVEL } from '../../permissions/types'
 import { ERROR_CODE, throwClientError } from '../../errors'
