+++ convex/gameMaps/functions/rollbackGameMap.ts
@@ -1,6 +1,6 @@
 import { asyncMap } from 'convex-helpers'
 import { ERROR_CODE, throwClientError } from '../../errors'
-import { requireItemAccess } from '../../sidebarItems/validation/validation'
+import { requireItemAccess } from '../../sidebarItems/validation/access'
 import { getSidebarItem } from '../../sidebarItems/functions/getSidebarItem'
 import { PERMISSION_LEVEL } from '../../permissions/types'
 import { logger } from '../../common/logger'
