+++ convex/gameMaps/functions/updateMap.ts
@@ -1,7 +1,5 @@
-import {
-  prepareSidebarItemRename,
-  requireItemAccess,
-} from '../../sidebarItems/validation/validation'
+import { prepareSidebarItemRename } from '../../sidebarItems/validation/orchestration'
+import { requireItemAccess } from '../../sidebarItems/validation/access'
 import { getSidebarItem } from '../../sidebarItems/functions/getSidebarItem'
 import { ERROR_CODE, throwClientError } from '../../errors'
 import { PERMISSION_LEVEL } from '../../permissions/types'
