+++ convex/notes/functions/createNote.ts
@@ -2,7 +2,7 @@ import * as Y from 'yjs'
 import { saveAllBlocksForNote } from '../../blocks/functions/saveAllBlocksForNote'
 import { syncNoteLinks } from '../../links/functions/syncNoteLinks'
 import type { ParsedCreateParentTarget } from '../../sidebarItems/validation/parent'
-import { prepareSidebarItemCreate } from '../../sidebarItems/validation/validation'
+import { prepareSidebarItemCreate } from '../../sidebarItems/validation/orchestration'
 import { resolveOrCreateFolderPath } from '../../folders/functions/resolveOrCreateFolderPath'
 import { SIDEBAR_ITEM_LOCATION, SIDEBAR_ITEM_TYPES } from '../../sidebarItems/types/baseTypes'
 import type { SidebarItemName } from '../../sidebarItems/validation/name'
