+++ convex/notes/functions/rollbackNote.ts
@@ -1,5 +1,5 @@
 import { rollbackYjsDocument } from '../../yjsSync/functions/rollbackYjsDocument'
-import { requireItemAccess } from '../../sidebarItems/validation/validation'
+import { requireItemAccess } from '../../sidebarItems/validation/access'
 import { getSidebarItem } from '../../sidebarItems/functions/getSidebarItem'
 import { PERMISSION_LEVEL } from '../../permissions/types'
 import { SIDEBAR_ITEM_TYPES } from '../../sidebarItems/types/baseTypes'
