+++ convex/sidebarItems/validation/parent.ts
@@ -1,7 +1,9 @@
 import { v } from 'convex/values'
 import type { Id } from '../../_generated/dataModel'
 import { ERROR_CODE, throwClientError } from '../../errors'
-import { requireSidebarItemName } from './name'
+import { SIDEBAR_ITEM_TYPES } from '../types/baseTypes'
+import { checkNameConflict, requireSidebarItemName, validateItemName } from './name'
+import type { AnySidebarItem } from '../types/types'
 import type { SidebarItemName, ValidationResult } from './name'
 
 export const CREATE_PARENT_TARGET_KIND = {
@@ -46,6 +48,23 @@ export const createItemParentArgsValidator = {
   parentTarget: createParentTargetValidator,
 } as const
 
+const VIRTUAL_PARENT = Symbol('virtual-parent')
+
+type ParentRef = Id<'sidebarItems'> | null | typeof VIRTUAL_PARENT
+type ParentLookup = { parentId: Id<'sidebarItems'> | null } | null | undefined
+type MaybePromise<T> = T | Promise<T>
+type ParentTargetValidationResult =
+  | {
+      valid: true
+      parentId: Id<'sidebarItems'> | null
+      siblings: Array<Pick<AnySidebarItem, '_id' | 'name'>>
+    }
+  | ({ valid: false } & Pick<Exclude<ValidationResult, { valid: true }>, 'error'>)
+
+function isPromiseLike<T>(value: MaybePromise<T>): value is Promise<T> {
+  return typeof value === 'object' && value !== null && 'then' in value
+}
+
 function requireParentPathSegment(segment: string): ParentPathSegment {
   const trimmedSegment = segment.trim()
 
@@ -74,11 +93,11 @@ export function requireCreateParentTarget(
   }
 }
 
-export function validateNoCircularParent(
+function validateNoCircularParentInternal(
   itemId: Id<'sidebarItems'>,
   newParentId: Id<'sidebarItems'> | null,
-  getParent: (id: Id<'sidebarItems'>) => { parentId: Id<'sidebarItems'> | null } | undefined,
-): ValidationResult {
+  getParent: (id: Id<'sidebarItems'>) => MaybePromise<ParentLookup>,
+): MaybePromise<ValidationResult> {
   if (!newParentId) {
     return { valid: true }
   }
@@ -91,11 +110,14 @@ export function validateNoCircularParent(
   }
 
   const seen = new Set<Id<'sidebarItems'>>()
-  let currentId: Id<'sidebarItems'> | null = newParentId
 
-  while (currentId) {
+  const visit = (currentId: Id<'sidebarItems'> | null): MaybePromise<ValidationResult> => {
+    if (!currentId) {
+      return { valid: true }
+    }
+
     if (seen.has(currentId)) {
-      break
+      return { valid: true }
     }
     seen.add(currentId)
 
@@ -107,8 +129,183 @@ export function validateNoCircularParent(
     }
 
     const current = getParent(currentId)
-    currentId = current?.parentId ?? null
+    if (isPromiseLike(current)) {
+      return current.then((value) => visit(value?.parentId ?? null))
+    }
+
+    return visit(current?.parentId ?? null)
   }
 
-  return { valid: true }
+  return visit(newParentId)
+}
+
+export function validateNoCircularParent(
+  itemId: Id<'sidebarItems'>,
+  newParentId: Id<'sidebarItems'> | null,
+  getParent: (id: Id<'sidebarItems'>) => ParentLookup,
+): ValidationResult {
+  const result = validateNoCircularParentInternal(itemId, newParentId, getParent)
+  if (isPromiseLike(result)) {
+    throw new Error('Invariant: synchronous parent lookup returned a Promise')
+  }
+
+  return result
+}
+
+export async function validateNoCircularParentAsync(
+  itemId: Id<'sidebarItems'>,
+  newParentId: Id<'sidebarItems'> | null,
+  getParent: (id: Id<'sidebarItems'>) => MaybePromise<ParentLookup>,
+): Promise<ValidationResult> {
+  return await validateNoCircularParentInternal(itemId, newParentId, getParent)
+}
+
+function buildParentStack(
+  parentId: Id<'sidebarItems'> | null,
+  itemsMap: Map<Id<'sidebarItems'>, AnySidebarItem>,
+): Array<Id<'sidebarItems'> | null> | null {
+  const stack: Array<Id<'sidebarItems'> | null> = [null]
+  if (parentId === null) return stack
+
+  const chain: Array<Id<'sidebarItems'>> = []
+  const seen = new Set<Id<'sidebarItems'>>()
+  let currentId: Id<'sidebarItems'> | null = parentId
+
+  while (currentId !== null) {
+    if (seen.has(currentId)) return null
+    seen.add(currentId)
+
+    const currentItem = itemsMap.get(currentId)
+    if (!currentItem) return null
+
+    chain.unshift(currentId)
+    currentId = currentItem.parentId
+  }
+
+  stack.push(...chain)
+  return stack
+}
+
+function findSidebarChildByName(
+  parentId: Id<'sidebarItems'> | null,
+  name: string,
+  parentItemsMap: Map<Id<'sidebarItems'> | null, Array<AnySidebarItem>>,
+): AnySidebarItem | undefined {
+  const normalizedName = name.trim().toLowerCase()
+
+  return parentItemsMap.get(parentId)?.find((item) => {
+    return item.name.trim().toLowerCase() === normalizedName
+  })
+}
+
+export function validateCreateParentTarget(
+  parentTarget: CreateParentTarget,
+  itemsMap: Map<Id<'sidebarItems'>, AnySidebarItem>,
+  parentItemsMap: Map<Id<'sidebarItems'> | null, Array<AnySidebarItem>>,
+): ParentTargetValidationResult {
+  if (parentTarget.kind === CREATE_PARENT_TARGET_KIND.direct) {
+    if (parentTarget.parentId !== null && !itemsMap.has(parentTarget.parentId)) {
+      return { valid: false, error: 'Parent not found' }
+    }
+
+    return {
+      valid: true,
+      parentId: parentTarget.parentId,
+      siblings: parentItemsMap.get(parentTarget.parentId) ?? [],
+    }
+  }
+
+  const parentStack = buildParentStack(parentTarget.baseParentId, itemsMap)
+  if (!parentStack) {
+    return { valid: false, error: 'Parent not found' }
+  }
+
+  const traversalStack: Array<ParentRef> = [...parentStack]
+
+  for (const segment of parentTarget.pathSegments) {
+    const trimmedSegment = segment.trim()
+
+    if (!trimmedSegment) {
+      return { valid: false, error: 'Path segments cannot be empty' }
+    }
+
+    if (trimmedSegment === '.') {
+      continue
+    }
+
+    if (trimmedSegment === '..') {
+      if (traversalStack.length === 1) {
+        return { valid: false, error: 'Path cannot traverse above the campaign root' }
+      }
+
+      traversalStack.pop()
+      continue
+    }
+
+    const nameResult = validateItemName(trimmedSegment)
+    if (!nameResult.valid) {
+      return nameResult
+    }
+
+    const currentParent = traversalStack[traversalStack.length - 1]
+    if (currentParent === VIRTUAL_PARENT) {
+      traversalStack.push(VIRTUAL_PARENT)
+      continue
+    }
+
+    const existingChild = findSidebarChildByName(currentParent, trimmedSegment, parentItemsMap)
+    if (!existingChild) {
+      traversalStack.push(VIRTUAL_PARENT)
+      continue
+    }
+
+    if (existingChild.type !== SIDEBAR_ITEM_TYPES.folders) {
+      return {
+        valid: false,
+        error: `"${trimmedSegment}" already exists here and is not a folder`,
+      }
+    }
+
+    traversalStack.push(existingChild._id)
+  }
+
+  const resolvedParent = traversalStack[traversalStack.length - 1]
+  if (resolvedParent === VIRTUAL_PARENT) {
+    return {
+      valid: true,
+      parentId: null,
+      siblings: [],
+    }
+  }
+
+  return {
+    valid: true,
+    parentId: resolvedParent,
+    siblings: parentItemsMap.get(resolvedParent) ?? [],
+  }
+}
+
+export function validateCreateItemLocally(
+  {
+    name,
+    parentTarget,
+  }: {
+    name: string
+    parentTarget: CreateParentTarget
+  },
+  itemsMap: Map<Id<'sidebarItems'>, AnySidebarItem>,
+  parentItemsMap: Map<Id<'sidebarItems'> | null, Array<AnySidebarItem>>,
+): ValidationResult {
+  const parentResult = validateCreateParentTarget(parentTarget, itemsMap, parentItemsMap)
+  if (!parentResult.valid) {
+    return parentResult
+  }
+
+  const trimmedName = name.trim()
+  const nameResult = validateItemName(trimmedName)
+  if (!nameResult.valid) {
+    return nameResult
+  }
+
+  return checkNameConflict(trimmedName, parentResult.siblings)
 }
